Move Repo conversion to MaybeOwnedRepo
This feels like a better place than on the Repo struct. This makes
MaybeOwnedRepo fully responsible for converting from string to Repo, and
keeps Repo concerned only with itself.

The other change is passing a Fn/Closure into the new conversion
function. Currently its very simple, but the tmux command will
eventually need some kind of function that takes in the name of the repo
and searches for that repo's presence (so that "neovim" as an argument
will find "neovim/neovim" instead of assuming "skipkayhil/neovim")