Fix t command with repos that contain dots
Tmux session names cannot contain a . or : because those characters are
used for attaching to windows and pane within windows:

  mysession:mywindow.1

Because of this, if a session is created with . or : in the name tmux
will replace it with an underscore automatically. This change fixes the
attach_cmd logic by applying the same replacement.