Refactor entrypoint around a Result
There were a couple places calling exit(1), which isn't great. Instead
of that, Command implementations should be returning Results. This also
allows error messages to be bubbled up in a nice way. Eventually that
could expand to more specific error messages, but for now a String was
the only thing being logged previously.

The Command implementations still have some work to do as they are
currently a mess still.