Add shell! macro, use it for simpler git clone
git::clone_cmd was doing too many different things. It created the
Command, called methods on Repo, and even looked at a field on the
config.

Instead of the command generation doing all this work, the work is
extracted to cmd::clone, and the command generation is replaced with a
basic macro.