Refactor tmux command to use new patterns
Specifically, don't pass a Repo around and use shell! to create
Commands. This removes the remaining references to Repo so the struct is
also removed.

A nice pattern was also found where Commands become very easy to print
when wrapped in PrintableCommand. If this continues to work well,
run_printable might end up disappearing as well.