Refactor find::repo to always deal with paths
Before, it would iterate directories but return the name of only the
owner. This required reconstructing the full path from the return value,
and made it hard to separate the case where the owner is specified in
the argument.

By making find::repo only deal with PathBufs, the code is dramatically
simpler and the owner-in-arg case is easily separated in
tmux::create_session (which allows the is_dir check to only be called in
that case, since find::repo will only return paths that have already
been checked).