Rewrite tmux command
Previously, the tmux command would allow creating or attaching to
a session based on a repo name argument. This was a great starting point
as it allowed me to quickly create a session named after a repo and the
working directory would already be the repo.

However, this workflow has a problem: the entire repo name must be typed
out every time its used. This resulted in me often using my tm script
to fuzzily attach to existing tmux sessions, as this would be faster
than typing out the whole repo name. In addition, this split between
tools meant that there is time spent thinking about which tool should be
used in the moment, which is really a waste of time.

To unify my workflow, sdev's tmux command now takes no arguments and
fuzzily creates/attaches to sessions. It pipes the list of local git
repos to fzf, and then based on the selection it either immediately
attaches or creates and then attaches to it.