Fix running tmux attach-session
Ref: b24a621f9703b19ca3d74b0eccfc5582e3413225

The tmux rewrite resulted in replacing run_printable with a simple match
on the output() of tmux attach-session. However, this is incorrect as
the tmux process needs to live until its eventually disconnected.

run_printable correctly used status(), and this commit fixes the tmux
attach-session match to use it as well