Refactor shell! to use PrintableCommand everywhere
Instead of removing run_printable, it ended up making sense to pull all
of the running/error handling/etc. into the PrintableCommand and then
using that everywhere. This completely removed the need for the
duplicate match + error handling for shell! commands. The simplification
even led to tmux::create_session becoming uneccesary since it ended up
just containing the shell! macro.

The only one missing is the fzf Command, however because it uses pipes
it will be a bit more complex to wrap.