Introduce Shell and Dep
Shell is meant to be a replacement for PrintableCommand now that more of
the program is coming together and I know a bit more rust. The first
improvement is that the macro no longer requires bringing Command into
scope because the macro refers to it by its absolute name. Additionally,
the macro is exported so that all of the logic can be moved out into its
own module. This becomes important as running shell commands will no
longer be in the cmd module.

The second addition is Dep (and its first implementation, Clone). This
is very much inspired by Babushka, which I found earlier this week. For
now all of the methods just return booleans, but they really need to
return Results for error handling. At this point I just want to start
off simple and as more Deps get implemented it should become clear
exactly what those Results will look like.