Make met return a Result, improve ShellError
This enables actually handling errors in Deps instead of treating them
the same as met returning false (they are similar in that they both mean
the Dep did not succeed but are very different in how they should be
displayed to the user)

Additionally, ShellError was improved a bit to keep the context of the
Error. Eventually this will probably end up with an anyhow::Result being
returned from commands so that error messages will contains a stack of
errors.