General Results improvements
This re-enables bubbling Results from Deps to Commands, which were
disabled for a bit. Additionally, it implements thiserror for more
errors so that bubbling will look nice. Finally, it adds anyhow::Result
as the return value for main() to automatically print the error stack
and exit with the proper exit status.