Add "open pr" command
Currently only implemented for a GitHub repository without an upstream
remote. This usually happens when you develop on a repository itself and
not a fork. Theoretically a repository could have no upstream and still
be a fork, but for my workflow I always setup upstream so that I can
keep my fork's base branches up to date.