Add fork support for "open pr"
This allows opening PRs from forks of repositories, meaning the PR is
opened against an upstream repo but the source branch is on the fork.

This also adds the opener crate to actually open the URL instead of just
printing it to the terminal. This is nice because it appears to work on
WSL (even though I don't expect it to work over ssh).