Fix Session Dep not normalizing session name
This caused `Session::met` to always return false, and so
`Session::meet` would always try to create the tmux session (which
errors when the session already exists).

To fix this, a new type was introduced to ensure that both of the Deps
must be used with a normalized session name. This doesn't quite seem
like the best layer to require the type, but since there isn't any
abstraction over tmux commands this seems like the best alternative.