Fix Dep::process succeeding on false second met
Previously, the return value of `::met` was returned as is from
`::process`. While this was correct for cases where calling `::met` the
second time would return an `Error`, it was not correct for cases where
it would return `Status::Unmet`. In those cases, `::process` should
actually have been returning an `Error` because `::met` returning false
means the `::meet` did not actually work.

This commit addresses the issue by changing the return type of
`::process` to `MeetResult`. This really just removes the `Status` from
the returned `Result`, so that there is no case where the `Result` can
be `Ok` but contain a `Status::Unmet`. This also cleaned up a bit of the
implementation of `::process` as well, since requirements would return
`UNMET` if their `::process` was `Unmet` when that case should also have
returned an `Error`.

Additionally, some very simple tests were added to exercise some of
these flows through `::process`. The mock structs ended up with a lot of
duplication but I haven't yet found a good way to implement defaults for
met/meet only inside the test config.