Fix unmet after meet warning for tmux command
Now that the final met of met/meet/met is properly raising an error, the
tmux command has started erroring on every invocation. This is due to
the final met running back in the parent process instead of inside the
tmux subprocess where it would evaluate to true.

So my solution for now is to inline the Attach Dep back into the tmux
command. It seems like it doesn't really fit the Dep model as well as
Session creation does.