Pass Path references around insted of PathBufs
There were a few awkward places where PathBufs had to be cloned that
didn't quite make sense (the most obvious one being where a PathBuf had
to be cloned because it was moved into try_from_absolute but may be
needed by the Error case as well). By passing around Path references we
can ensure that the only clones are explicitly when we want to put a
PathBuf in a struct.