Refactor GitRepo try_from_*
There are two things being changed here:

The first change is removing the Path value from NotARepo because it is
no longer needed. It was previously necessary because the error case of
try_from_fs needed to read from the PathBuf moved into the function, but
now that the param is a Path reference nothing is being moved.

The second set of changes is replacing a lot of ifs with Option <->
Result methods. These prevent needing all of the "return Err" type
expressions which seem overly verbose.