Add experimental z command
z is going to be a playground for testing things. In this case, z now
implements the easy part of rendering an fzf like ui for fuzzy matching.
In a future commit the (probably) harder part of actually doing matching
will be added, and the whole thing will eventually replace the tmux
command.