Add custom selected to App struct
In the interest of replacing the standard List with a custom one that
more accurately represents the possible states of the program, this
commit adds a custom selected index to the App struct. It uses u32 to
match the size of Nucleo's match list, which makes it easier to lookup
the final match on completion.

Additionally, this patch also fixes a bug where the selected row would
be hidden when the number of matches drops below the selected index.
Previously that case was not handled and the selected row would be
hidden above the matched list, but now the selected row will always stay
within the range of matches.