Remove host/GitHost from GitRepo
In preparation for some refactors around the directory traversal to
build the list of GitRepos, its time to finally remove the unused host
field. The original intention was for GitRepo to be shared between
multiple commands, but it still has yet to be used outside of building
the list of paths for the tmux command.