Refactor Unmeetable to wrap any errors
If Dep were ever to be extracted to a library, the `Unmeetable` error
needs to be able to be `try!/?`d with any error type.

This commit enables that by aliasing `Unmeetable` as `anyhow::error`,
which will prevent needing to enumerate all of the possible errors that
could happen inside a `met`/`meet` (which is theoretically any error).