Refactor clone to always go through gix::Url
The Clone Dep will need a gix::Url to start cloning repositories using
gix.

This refactor prepares for that future state by first constructing a gix::Url
from GitRepoSource instead of going straight to String. The conversion
to String is still needed for shell!, but once shell! is replaced the
String conversion will also be removed.