Hide all common ancestors from repo paths
Previously the paths were displayed relative to the "src root", however
it may be common for all repo paths to use the same host and even the
same path under the host.

This commit calculates the longest path prefix while adding repos to
Nucleo so that the paths have the potential to be shorter. If multiple
hosts (or paths under a single host) are present, this change may have
no effect.